HUDSON-6551: improved Doxygen parser to support almost any warning it emits, and retrieve multiline messages.
I had to discard the content of the old test file doxygen.txt, because there were relative paths. I was not able to reproduce such a report: whatever the platform and the settings, Doxygen always print absolute paths.
As relative path are incompatible with my parser, and would be considerably harder to support (while keeping multiline support), I chose to ignore this issue.