Improved the memory consumption at runtime by selective String.intern.
I have a 28MB findbugs-warnings.xml that contains 23674 bugs, which
occupies 40.1MB in heap when read. Origin, category, and type are all
enum-like strings, and so by interning it, I cut down the heap
consumption down to 33.5MB, a 16% saving.

String.intern() puts pressure on permgen, so there's a risk in overusing
it, but I think it's pretty safe to use this for those limited things.