ParserResult: Fix checking for duplicate annotations
Previously, if an annotation is not yet contained in the set, its path gets
expanded and then added. This leads to subsequent checks for duplicates to
fail because the annotation in the set has the path expanded whereas the
incoming duplicate still has the original path.

Fix this by always expanding the path first. Except for duplicates we need
to do this anyway, and given the assumption that there are not many
duplicates we are not calling expandRelativePaths(), which might be quite
expensive, much more than before.

As an added benefit this only checks once for the annotation already being
contained in the set. Before it was checked twice, once explicitly by
contains(), and then again implicitly by add().