Set a arbitrary limit to the warning message length
Found that when there are a lot of identical stacktrace in the Maven log without any
blank lines between them, the condensed warning list concatenate all the message into
one. This provoques an OOM error due to the `veryLongString + "\n" + shorterString`
code.

This should only be considered as a hotfix. Maven warning parsing process should be
review to better manipulate the warning detections when there is no blank line to
separate them.