Clarify the wording about reference builds
Make more clear that there is no such thing as a reference build and
stable / previous builds, but that the reference build always refers to
either stable / previous builds.

Also reorder the sections slightly to resemble the order in the UI and
consistently use the term "compute" in favor of "calculate".