add Linux Kernel Output Parser
Parse Linux Kernel Output for BUG() and WARN() output. Notice all other
Output with priority LOW.

BUG()/WARN() create additional information about the failure. These
information depend on boot, CPU and other runtime conditions.

To be able to track the failure in new, persist and fixed add this
additional inforamtion into the tooltip. The Information is available to
the user, but do not disturb the failure tracking.

The plugin is used in the CI-RT Testframework to recognize failures on
target test systems.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Holger Dengler <dengler@linutronix.de>