[JENKINS-75344] Fix custom URL name of result action
In https://github.com/jenkinsci/warnings-ng-plugin/pull/1919
a new icon property has been added to the recorder.
This accidentally broke the URL mapping of the warning results.

When a custom ID is given, then this ID will be used as URL.
Only for empty URLs, the default ID of the label provider will be used.

Additionally, the log messages have been fixed when the ID
is used in tool and recorder (see JENKINS-75391).