[JENKINS-75647] Use only 5 builds when scanning the history for actions
The `MissingResultFallbackHandler` scans the build history of a job
for `ResultAction`s to provide trend charts even for jobs that have no
valid result in the last build. It makes sense to look only at the last
couple of builds for such actions. Otherwise, too many builds need to be
serialized in the Jenkins controller: this will lead to performance
problems.

Fixes https://issues.jenkins.io/browse/JENKINS-75647